New database script for Nathanos Blightcaller.
Better than the old hardcoded script! Now he says all his texts and the whole skull fetching event is implemented, and its done properly with the spells Blizzard used. The logic for the skeleton spawning spell in the old script was completely wrong, it is not used when somebody heals, its just a timer but the chance is not 100% so some people tried to attribute a reason to the randomness. Multiple comments disprove it, and you can see that in the warrior solo video linked in the migration he summons a ton of skeletons, even though the warrior never heals.